United States Farm Service Agency Acreage Data

This package contains annual county level data from the United States Farm Service Agency.

Acreage data from 2009-2023 by County, irrigation, and other factors.

Usage

acresFSA

Format

A data frame with nine variables:

Year

Year, 2009-2023

State, County

US State and county names, all caps.

Type

Cultivation type: Failed, Not Planted, Planted, Prevented, or Volunteer

Irrigation Practice

Irrigated with typical means, other type of irrigation, and total (irrigated and non-irrigated)

Crop

Name of crop

Variety

Name of crop variety

Intended Use

Intended use of the crop

Acres

Amount of acres

For further details, see https://www.fsa.usda.gov/news-room/efoia/electronic-reading-room/frequently-requested-information/crop-acreage-data/index

See also

Author

Maintainer: C A Pellett c.alex.pellett@gmail.com

Other contributors:

State of South Carolina Department of Natural Resources [funder]